+++ src/protocol/message_process.h
@@ -12,5 +12,9 @@ void message_process_handler(void *argv);
 void get_readings_timer_callback(void *);
 #endif
 
+void send_address_by_sms();
+
+void send_readings_by_sms(char *message);
+
 #endif
 #endif