+++ test/test.cpp
@@ -4,18 +4,16 @@
 #include <stdio.h>
 #include <stdint.h>
 #include <string.h>
+#include <stdlib.h>
+#include <stdbool.h>
 
+#include "commands_test.h"
+#include "minunit.h"
 /******************
  * INCLUDES TO TEST
  ******************/
 #include "../src/protocol/protocol.h"
 
-/******************
- * DEFINES
- ******************/
-#define mu_assert(message, test) do { if (!(test)) return message; } while (0)
-#define mu_run_test(test) do { char *message = test(); tests_run++; \
-                                if (message) return message; } while (0)
 
 /******************
  * GLOBAL VARIABLES
@@ -104,6 +102,52 @@ static char * all_tests()
     return 0;
 }
 
+static bool command_validate_phone_number(const char *phone)
+{
+    bool status = false;
+    int digits = 0;
+    char number[15];
+    printf( "Verificando se %s é um telefone válido...\n", phone);
+    sscanf(phone, " %11[0-9] %n", number, &digits);
+    /*
+    if (digits == 0)  // no digits
+    {
+        printf("%s - sem dígitos\n", __func__);
+        return false;
+    }
+    if (number[digits]) // junk after the digits
+    {
+        printf("%s - dígitos extras\n", __func__);
+        return false;
+    }
+    if (strlen(number) == 11 || strlen(number) == 9)
+    {
+        return true;
+    }
+    else
+    {
+        printf("%s - não tem o tamanho de um número\n", __func__);
+        return false; // not 7 digits
+    }*/
+
+    if (digits == 0 && number[digits])  // no digits
+    {
+        status = false;
+    }
+    else
+    {
+        if (strlen(number) == 11 || strlen(number) == 9)
+        {
+            status = true;
+        }
+        else
+        {
+            status = false;
+        }
+    }
+    
+    return status;
+}
 /******************
  * MAIN FUNCTION
  ******************/
@@ -121,6 +165,19 @@ int main(void)
         printf("ALL TESTS PASSED\n");
     }
     printf("Tests run: %d\n", tests_run);
+    test_commands();
 
+    printf("telefone: %s\n", 
+            command_validate_phone_number("04141998271302")?"True":"False");
+    printf("telefone: %s\n", 
+            command_validate_phone_number("041998271302")?"True":"False");
+    printf("telefone: %s\n", 
+            command_validate_phone_number("998271302")?"True":"False");
+    printf("telefone: %s\n", 
+            command_validate_phone_number("+5541998271302")?"True":"False");
+    printf("telefone: %s\n", 
+            command_validate_phone_number("ok")?"True":"False");
+
+    printf("%d\n", (int)strtol("intervalo 30", (char**)NULL, 10));
     return result != 0;
 }