+++ CHANGELOG.md
@@ -28,6 +28,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `sim7070g_read_sms_by_id` function to get sms stored on device by `ID`;
 * Added `sim7070g_list_all_sms` function to get all sms stored on device;
 * Added `sim7070g_clear_sms_list` function delete all sms stored on device;
+* Added `commands` module to parse SMS commands;
+* Added `commands_utils` module to help with the commands;
+* Added task to handle the process of sending sms to `sim7070g`;  
+* Added command response to `command_handler`. Now it responds the phone number with the result of the command;
 
 ### Changed